refactor: 커스텀 hook 안의 데이터 요청을 request 커스텀 훅을 통해 일괄 처리
