+++ src/hooks/orderDetail.ts
@@ -1,30 +1,17 @@
-import { useEffect, useState } from 'react';
+import { useState } from 'react';
 import { requestGetOrder } from '../apis/order';
 import { Order } from '../type';
 import { parseOrderData } from '../utils/parseData';
+import useRequest from './request';
 
 const useOrderDetail = (orderId: Order['id']) => {
   const [orderItems, setOrderItems] = useState<Order['orderItems']>([]);
-  const [loading, setLoading] = useState<boolean>(true);
-  const [responseOK, setResponseOK] = useState<boolean>(true);
 
-  useEffect(() => {
-    const fetchData = async () => {
-      setLoading(true);
-      try {
-        const response = await requestGetOrder(orderId);
-        const order = response.data;
-        setOrderItems(parseOrderData(order).orderItems);
-        setResponseOK(true);
-      } catch (error) {
-        console.error(error);
-        setResponseOK(false);
-      } finally {
-        setLoading(false);
-      }
-    };
-    fetchData();
-  }, [orderId]);
+  const { loading, responseOK } = useRequest(async () => {
+    const response = await requestGetOrder(orderId);
+    const order = response.data;
+    setOrderItems(parseOrderData(order).orderItems);
+  });
 
   return { orderItems, loading, responseOK };
 };
